Implement JSONPath (#8)
* Implement JSONPath

Copy files from https://github.com/markborkum/pacifica-jsonpath@jsonpath2 and merge conflicts.

* Fix some Flake8 warnings

* Initial pre-commit run

Signed-off-by: David Brown <dmlb2000@gmail.com>

* Test parse_str method.

* Operator callables should be static

Flake8 warned that the operator callables were globals. In a previous commit, the operator callables were refactored into method-locals in the operator's constructor. This is sufficient to resolve the flake8 warnings, but introduces a bug: since callables are created per instance, they have different memory locations, and hence, are not compatible with the `Node`-level definition of `__eq__`, which naively compares each `__dict__` of each operand. The fix is to refactor the operator callables into static methods.

* pre-commit fixes

* add some bookstore examples for testing

Signed-off-by: David Brown <dmlb2000@gmail.com>

* pre-commit fixes

Signed-off-by: David Brown <dmlb2000@gmail.com>

* Rollback modification to grammar

* Fix bookstore tests

In previous implementations of JSONPath, the current value is implicitly cast to array or object if the next node is an array index subscript or object index subscript. This introduces ambiguity: "is the index subscript referring to the current value, or a child value of the current value?" The solution is to use the wildcard "*" to refer to the child values explicitly.

* add more testing to increase coverage

Signed-off-by: David Brown <dmlb2000@gmail.com>

* pre-commit fix

Signed-off-by: David Brown <dmlb2000@gmail.com>

* more testing coverage

* Improve slice notation

Aim for feature parity with equivalent in Python programming language

* Tests to increase coverage

* pre-commit fixes

* fix pre-commit

* try some more non-sense tests

* add array to test

* Python 3.5 doesn't seem to be working great

* Add coverage pragmas

* More coverage tests

* try getting pre-commit and testing right